Update bamazonCustomer-promise.js
Problem was in that your original function to displayAllItems was not wrapped in a promise. Here's the full breakdown

You would connect then return a promise, that promise then executed displayAllItems which ran connection.query synchronously and returned nothing. The next promise then executed. 

To stop that from happening and for connection query to be looped into the main promise chain, it needed to be promisified and then returned to the chain. Once you return a promise to the original promise chain, it 'knows' that it needs to wait for the promisified connection.query to return before proceeding.